[core] now using Index instead of std::size_t (compatibility with Eigen::Index)

+++ src/core/actions/codac2_OctaSym.cpp
@@ -20,7 +20,7 @@ OctaSym::OctaSym(const vector<int>& s)
 {
   for(const auto& i : s)
   {
-    assert_release((size_t)std::abs(i) > 0 && (size_t)std::abs(i) <= size());
+    assert_release(std::abs(i) > 0 && std::abs(i) <= size());
   }
 }
